Added new samples & fixed some issues
- Added Adm2Txt & Txt2Adm
- Fixed an issue with the HashUtility
- Extended GUIDgen with a copy button

+++ src/Utilities/HashUtility.cs
@@ -7,7 +7,7 @@
             int hash = str.Length;
             foreach (var c in str)
             {
-                hash = (hash >> 27) ^ (hash << 5) ^ c;
+                hash = (int)((uint)hash >> 27) ^ (hash << 5) ^ c;
             }
 
             return hash;
